+++ packages/expo-sensors/android/src/main/java/expo/modules/sensors/modules/PedometerModule.java
@@ -9,8 +9,9 @@ import android.os.Bundle;
 
 import org.unimodules.core.Promise;
 import org.unimodules.core.interfaces.ExpoMethod;
-import org.unimodules.interfaces.sensors.SensorService;
-import org.unimodules.interfaces.sensors.services.PedometerService;
+
+import expo.modules.interfaces.sensors.SensorServiceInterface;
+import expo.modules.interfaces.sensors.services.PedometerServiceInterface;
 
 public class PedometerModule extends BaseSensorModule {
   private Integer stepsAtTheBeginning = null;
@@ -30,8 +31,8 @@ public class PedometerModule extends BaseSensorModule {
   }
 
   @Override
-  protected SensorService getSensorService() {
-    return getModuleRegistry().getModule(PedometerService.class);
+  protected SensorServiceInterface getSensorService() {
+    return getModuleRegistry().getModule(PedometerServiceInterface.class);
   }
 
   protected Bundle eventToMap(SensorEvent sensorEvent) {