+++ packages/expo-sensors/android/src/main/java/expo/modules/sensors/services/AccelerometerService.java
@@ -10,7 +10,9 @@ import java.util.List;
 
 import org.unimodules.core.interfaces.InternalModule;
 
-public class AccelerometerService extends SubscribableSensorService implements InternalModule, org.unimodules.interfaces.sensors.services.AccelerometerService {
+import expo.modules.interfaces.sensors.services.AccelerometerServiceInterface;
+
+public class AccelerometerService extends SubscribableSensorService implements InternalModule, AccelerometerServiceInterface {
   public AccelerometerService(Context reactContext) {
     super(reactContext);
   }
@@ -22,6 +24,6 @@ public class AccelerometerService extends SubscribableSensorService implements I
 
   @Override
   public List<Class> getExportedInterfaces() {
-    return Collections.<Class>singletonList(org.unimodules.interfaces.sensors.services.AccelerometerService.class);
+    return Collections.<Class>singletonList(AccelerometerServiceInterface.class);
   }
 }