+++ packages/expo-modules-core/CHANGELOG.md
@@ -8,6 +8,10 @@
 
 ### 🐛 Bug fixes
 
+### 💡 Others
+
+- Added `Interface` suffix to sensors interfaces names to get rid of name collisions. ([#12888](https://github.com/expo/expo/pull/12888) by [@tsapeta](https://github.com/tsapeta))
+
 ## 0.0.1 — 2021-05-07
 
 _This version does not introduce any user-facing changes._